requirements.txt
Django~=3.2.16
django-environ~=0.9.0
django-recaptcha2~=1.4.1
django-widget-tweaks~=1.4.12
django-bootstrap4~=22.2
django-bootstrap-datepicker-plus~=5.0.0
python-dateutil~=2.8.2
gunicorn~=20.1.0
whitenoise~=6.2.0
psycopg2~=2.9.5
line-bot-sdk~=2.4.2
openai~=0.27.5